module_3(fix): rename CreditCardPage and its components into LoanPage
