--- src/components/page-blocks/credit-card-page-blocks/banner-credit-card/index.ts
@@ -1,1 +1,0 @@
-export * from "./banner-credit-card";
