--- src/components/page-blocks/credit-card-page-blocks/credit-card-tabs/credit-card-tabs.tsx
@@ -1,6 +1,0 @@
-import { Tabs } from "@components/ui/tabs";
-import { CREDIT_CARD_TABS_ITEMS } from "./credit-card-tabs-items.const";
-
-export const CreditCardTabs = () => {
-  return <Tabs tabs={CREDIT_CARD_TABS_ITEMS} />;
-};
