--- src/components/page-blocks/credit-card-page-blocks/credit-card-tabs/index.ts
@@ -1,2 +1,0 @@
-export * from "./credit-card-tabs";
-export * from "./credit-card-tabs-items.const";
