--- src/components/page-blocks/credit-card-page-blocks/index.ts
@@ -1,4 +1,0 @@
-export { GetCard } from "./get-card";
-export { BannerCreditCard } from "./banner-credit-card";
-export { CreditCardTabs } from "./credit-card-tabs";
-export { Form, FormTitle, PrescoringForm } from "./form";
