+++ src/components/page-blocks/loan-page-blocks/banner-loan/banner-loan.module.scss
