+++ src/components/page-blocks/loan-page-blocks/banner-loan/banner-loan.tsx
@@ -2,14 +2,14 @@ import { Button } from "@components/ui/button";
 import { CreditCard } from "@components/ui/credit-card";
 import { InfoBlock } from "@components/ui/info-block/info-block";
 import { Tooltip } from "@components/ui/tooltip";
-import styles from "./banner-credit-card.module.scss";
-import { BANNER_CREDIT_CARD_ITEMS } from "./banner-credit-card-items.const";
+import { BANNER_LOAN_ITEMS } from "./banner-loan-items.const";
+import styles from "./banner-loan.module.scss";
 
-export type BannerCreditCardPropsT = {
+export type BannerLoanPropsT = {
   handleScroll: () => void;
 };
 
-export const BannerCreditCard = ({ handleScroll }: BannerCreditCardPropsT) => {
+export const BannerLoan = ({ handleScroll }: BannerLoanPropsT) => {
   return (
     <section role="banner" className={styles.banner__wrapper}>
       <h1 className={styles.banner__heading}>Platinum digital credit card</h1>
@@ -19,7 +19,7 @@ export const BannerCreditCard = ({ handleScroll }: BannerCreditCardPropsT) => {
       </p>
       <div className={styles.infoBlock__wrapper}>
         <ul className={styles.infoBlock__list}>
-          {BANNER_CREDIT_CARD_ITEMS.map(item => (
+          {BANNER_LOAN_ITEMS.map(item => (
             <li key={item.id}>
               <Tooltip text={item.tooltip}>
                 <InfoBlock
