+++ src/components/page-blocks/loan-page-blocks/banner-loan/index.ts
@@ -1,0 +1,1 @@
+export * from "./banner-loan";
