+++ src/components/page-blocks/loan-page-blocks/get-card/get-card.tsx
