+++ src/components/page-blocks/loan-page-blocks/get-card/info-block/info-block.module.scss
