+++ src/components/page-blocks/loan-page-blocks/index.ts
@@ -1,0 +1,5 @@
+export { GetCard } from "./get-card";
+export { BannerLoan } from "./banner-loan";
+export { LoanTabs } from "./loan-tabs";
+export type { LoanTabT } from "./loan-tabs/";
+export { Form, FormTitle, PrescoringForm } from "./form";
