+++ src/components/page-blocks/loan-page-blocks/loan-tabs/components/about-card/about-card.tsx
