+++ src/components/page-blocks/loan-page-blocks/loan-tabs/components/cashback/cashback-items.const.ts
