+++ src/components/page-blocks/loan-page-blocks/loan-tabs/components/rates-conditions/rates-conditions-items.const.ts
