+++ src/components/page-blocks/loan-page-blocks/loan-tabs/index.ts
@@ -1,0 +1,3 @@
+export * from "./loan-tabs";
+export * from "./loan-tabs-items.const";
+export type { LoanTabT } from "./loan-tabs-items.const";
