+++ src/pages/index.ts
@@ -2,6 +2,6 @@ export { ResourcesPage } from "@pages/resources-page/resources-page";
 export { ProductPage } from "@pages/product-page/product-page";
 export { Page404 } from "@pages/page-404/page-404";
 export { MainLayoutPage } from "@pages/main-layout-page/main-layout-page";
-export { CreditCardPage } from "./creadit-card-page/credit-card-page";
+export { LoanPage } from "./loan-page/loan-page";
 export { AccountPage } from "./account-page/account-page";
 export { HomePage } from "@pages/home-page/home-page";
